Implement API subroutines and global module

+++ pyfvs/fvs.py
@@ -9,8 +9,11 @@ Created on Nov 29, 2014
 @author: tod.haren@gmail.com
 """
 
+from dataclasses import field
+from distutils.log import debug
 import os
 import sys
+import re
 import shutil
 import logging
 import logging.config
@@ -45,7 +48,7 @@ class FVS(object):
 
     Access to the FVS API routines as well as additional FVS core routines,
         and common arrays and variables are accessible using standard Python
-        attribute access conventions, eg. cycles = fvs.contrl.ncyc to find out
+        attribute access conventions, eg. cycles = fvs.globals.ncyc to find out
         how many projection cycles were executed. API and other subroutines
         have expected input variables and return numpy arrays or Python objects.
 
@@ -57,11 +60,10 @@ class FVS(object):
     Attributes:
         variant: FVS variant abbreviation, PN, WC, etc. (Required)
         stochastic: If True the FVS random number generater will be
-                reseeded on each call to run_fvs. If False the generator
-                will be set to a fixed value of 1.0. (Optional)
+                reseeded on each run. If False the generator will
+                be set to a fixed value of 1.0. (Optional)
         bootstrap: Number of bootstrap resamples of the plot data to
                 execute. (Optional)
-        config: Configuration dictionary. (Optional)
 
     Usage:
         fvs = FVS(<variant abbreviation>)
@@ -70,7 +72,7 @@ class FVS(object):
        # Return a tuple of species codes given an FVS ID using the FVS API.
        spp_attrs = fvs.fvsspeciescode(16)
     """
-    def __init__(self, variant, stochastic=False, bootstrap=0, config=None, cleanup=True, treelist_format=None
+    def __init__(self, variant, stochastic=False, bootstrap=0, cleanup=True, treelist_format=None
             , workspace=None):
         """
         Initialize a FVS variant library.
@@ -81,30 +83,44 @@ class FVS(object):
         self.stochastic = stochastic
         self.bootstrap = bootstrap
         self.cleanup = cleanup
-        self._workspace = workspace
+        self.workspace = workspace
 
         self.root = None
         self.keywords_file = None
-        
-        if not config:
-            self.config = pyfvs.get_config()
-        else:
-            self.config = config
+
+        # Resolve the confuse configuration files
+        self.config = pyfvs.config
 
         self._random_seed = 55329  # Default FVS random number seed
 
         self.fvslib_path = None
         self.fvslib = None
-        
+
+        if self._workspace is None:
+            self._workspace = self.config['workspace'].get()
+
         if treelist_format is None:
             # Use the default PyFVS treelist template
             self.treelist_tmplt = pyfvs.treelist_format['template']
             self.treelist_fmt = pyfvs.treelist_format['fvs_format']
-        
+
         else:
             self.treelist_tmplt = treelist_format['template']
             self.treelist_fmt = treelist_format['fvs_format']
-            
+
+        self._tree_fields = re.findall('{(.+?):.*?}', self.treelist_tmplt)
+        self._inv_tree_defaults = dict(
+            tree_history=1, dg_incr=0.0, live_ht=0.0
+            , trunc_ht=0.0, htg_incr=0.0, crown_ratio=0
+            , damage1=0, severity1=0
+            , damage2=0, severity2=0
+            , damage3=0, severity3=0
+            , tree_val=1, tree_rx=1
+            , plot_slope=0, plot_aspect=0, plot_habitat=0
+            , plot_topo_pos=0, plot_site_prep=0
+            , tree_age=0.0
+        )
+
         self._load_fvslib()
 
         # if self.fvslib is None:
@@ -112,7 +128,7 @@ class FVS(object):
 
         self._keywords = None
         # self.fvs_trees = FVSTrees(self)
-        self._inv_trees = None
+        self._inventory_trees = None
         self._spp_codes = None
         self._spp_seq = None
         self._spp_fia_codes = None
@@ -121,41 +137,103 @@ class FVS(object):
         # Keep a list of files and folders to cleanup
         self._artifacts = []
 
+        # print(self.workspace)
+
+        self.version_info = dict(
+            variant = self.variant
+            , comp_date = self.fvslib.globals.compile_date.astype(str)
+            , comp_time = self.fvslib.globals.compile_time.astype(str)
+            , rev_tag = self.fvslib.globals.svn.astype(str)
+            , rev_date = self.fvslib.revise(self.variant).strip().decode()
+            )
+
+
     # @property
     # def trees(self):
-    #     deprecation('FVS.trees is deprecated. Use FVS.fvs_trees or FVS.inv_trees')
+    #     deprecation('FVS.trees is deprecated. Use FVS.fvs_trees or FVS.inventory_trees')
     #     return self.fvs_trees
-    
+
     @property
-    def inv_trees(self):
+    def inventory_trees(self):
         """
         A dataframe of inventory tree records.
-        
+
         Columns correspond to the description in Essential FVS section 4.2:
             (plot_id,tree_id,prob,tree_history,species,dbh,dg_incr
                 ,live_ht,trunc_ht,htg_incr,crown_ratio,
                 ,dmg1,svr1,dmg2,svr2,dmg3,svr3
                 ,value_class
         """
-        
-        return self._inv_trees
-    
-    @inv_trees.setter
-    def inv_trees(self, trees):
+
+        return self._inventory_trees
+
+    @inventory_trees.setter
+    def inventory_trees(self, trees):
         """
         Set the inventory trees dataframe
         """
-        
-        # check for required fields
-        rqd = ['plot_id','tree_id','prob','species','dbh']
-        missing = [fld for fld in rqd if not fld in trees.columns]
-        if missing:
-            raise ValueError(
-                    'Inventory trees missing required fields: '
-                    '{}'.format(str(missing)))
-                
-        self._inv_trees = trees
-       
+
+        ## TODO: Check for required fields
+        ## TODO: Defer loading arrays until the simulation is started
+
+        ## FVS array : Input field
+        field_map = dict(
+            plot_id='plot_id'
+            , tree_id='tree_id'
+            , history='history'
+            , prob='prob'
+            , species='species'
+            , diameter='diameter'
+            , diameter_growth='diameter_growth'
+            , height='height'
+            , total_height='total_height'
+            , height_growth='height_growth'
+            , crown='crown'
+            , age='age'
+        )
+
+        trees = trees.loc[:,field_map.values()].copy()
+        trees.rename(columns={v:k for k,v in field_map.items()}, inplace=True)
+
+        # Now load the arrays
+        inv = self.fvslib.inventory_trees
+        nrows = trees.shape[0]
+        inv.initialize(nrows)
+
+        inv.plot_id = trees['plot_id']
+        inv.tree_id = trees['tree_id'] #.astype(int)
+
+        inv.tree_status = trees['history']
+
+        inv.trees = trees['prob']
+        inv.species[:nrows] = trees['species']
+        inv.diameter = trees['diameter']
+        inv.diameter_growth = trees['diameter_growth']
+        inv.height = trees['height']
+
+        # # Broken and dead tops have actualht<height
+        # tht = np.zeros(n)
+        # m = plot_trees['actualht']<plot_trees['ht']
+        # tht[m] = plot_trees['actualht'][m]
+        # inv.total_height = tht
+
+        inv.height_growth = trees['height_growth']
+        inv.crown_ratio = trees['crown']
+        inv.tree_age = trees['age']
+
+        ## TODO: Add damage/severity columns
+
+        # # check for required fields
+        # rqd = ['plot_id','tree_id','tree_count','species','diameter']
+        # missing = [fld for fld in rqd if not fld in trees.columns]
+        # if missing:
+        #     raise ValueError(
+        #             'Inventory trees missing required fields: '
+        #             '{}'.format(str(missing)))
+
+        # print(trees)
+        self._inventory_trees = trees
+
     def _load_fvslib(self):
         """
         Load the requested FVS variant library.
@@ -204,7 +282,7 @@ class FVS(object):
             self._spp_seq = seq
 
         return self._spp_seq
-    
+
     @property
     def spp_fia_codes(self):
         """
@@ -226,18 +304,26 @@ class FVS(object):
             self._spp_plants_codes = dict(zip(self.spp_codes,jsp))
 
         return self._spp_plants_codes
-    
+
     def __getattr__(self, attr):
         """
-        Return an attribute from self.fvslib if it is n ot defined locally.
+        Return an attribute from self.fvslib if it is not defined locally.
         """
+        if self.fvslib is None:
+            # return None
+            raise AttributeError
+
         try:
-            return getattr(self.fvslib, attr)
+            v = getattr(self.fvslib, attr, None)
+            if v is None:
+                v = getattr(self.fvslib.globals, attr)
 
         except AttributeError:
-            # msg = 'No FVS object {}.'.format(attr,)
+            msg = 'No FVS object "{}".'.format(attr,)
             # log.exception(msg)
             raise AttributeError(msg)
+
+        return v
 
     def set_random_seed(self, seed=None):
         """
@@ -263,7 +349,7 @@ class FVS(object):
     def _init_fvs(self, keywords):
         """
         Initialize FVS with the given keywords file.
-        
+
         Args:
             keywords: Path of the keyword file initialize FVS with.
         """
@@ -281,10 +367,10 @@ class FVS(object):
         """
         Return the current workspace folder.
         """
+        pfx = self.config['prefix'].get()
         if (self._workspace is None
                 or not os.path.exists(self._workspace)):
-            # FIXME: look in self.config first
-            self._workspace = tempfile.mkdtemp(prefix='pyfvs')
+            self._workspace = tempfile.mkdtemp(prefix=pfx)
             self._artifacts.append(self._workspace)
 
         return self._workspace
@@ -293,11 +379,11 @@ class FVS(object):
     def workspace(self, workspace):
         """
         Set the current workspace folder.
-        
+
         Args:
             workspace: Path to set as the current FVS workspace.
         """
-        if not os.path.exists(workspace):
+        if not workspace is None and not os.path.exists(workspace):
             raise ValueError('Workspace folder must already exist.')
 
         self._workspace = workspace
@@ -319,7 +405,7 @@ class FVS(object):
     def keywords(self, keywords):
         """
         Set the current keywords instance.
-        
+
         Args:
             keywords: Instance of KeywordSet class.
         """
@@ -332,11 +418,11 @@ class FVS(object):
     def init_keywords(self, title='', comment=''):
         """
         Return an initialized KeywordSet instance.
-        
+
         Args:
             title: Keywords title
             comment: Keywords comment
-        
+
         Returns:
             keywords: An initialized KeywordSet object
         """
@@ -344,7 +430,7 @@ class FVS(object):
             now = datetime.datetime.strftime(
                     datetime.datetime.now(), '%Y%m%d_%H%M%S')
             title = '{}_{}'.format(self.variant,now)
-            
+
         self._keywords = kw.KeywordSet(
                 title=title, comment=comment, top_level=True
                 )
@@ -353,7 +439,7 @@ class FVS(object):
     def init_projection(self, keywords=None, trees=None):
         """
         Initialize a projection with the provided keywords.
-        
+
         Args:
             keywords: Path to the FVS keywords file, or a KeywordSet instance.
                     If None then use self.keywords.
@@ -363,22 +449,25 @@ class FVS(object):
         if keywords is None:
             keywords = self.keywords
 
+        if keywords is None:
+            print('Run has no Keywords')
+
         if isinstance(keywords, kw.KeywordSet):
-            
+
             # FIXME: keyword implicitly know their relative position
             keywords.top_level = True
             keywords.parent = None
 
             # TODO: Add option to raise exceptions for missing keywords
-            if not keywords.find('STDINFO'):
+            if not keywords.find('STDINFO') and not keywords.find('STANDSQL'):
                 print('No STDINFO keyword')
-                        
+
             if not keywords.find('DESIGN') and not keywords.find('STANDSQL'):
                 print('No DESIGN keyword')
-                
+
             if not keywords.find('TREEFMT') and not keywords.find('TREESQL'):
                 keywords += kw.TREEFMT(self.treelist_fmt)
-                
+
             fn = keywords.title.lower()
             fn = fn.replace(' ', '_')
             self.root = os.path.join(self.workspace, fn)
@@ -389,12 +478,14 @@ class FVS(object):
                     self.root
                     , '{}.key'.format(fn)
                     )
+            log.debug(f'Keywords File: {keywords_fn}')
             keywords.write(keywords_fn)
 
             self._artifacts.append(keywords_fn)
 
             # Write out the inventory trees
-            if not keywords.find('TREESQL'):
+            has_inv = self.fvslib.inventory_trees.row_count()>0
+            if not has_inv and not keywords.find('TREESQL'):
                 # print('***NO TREESQL')
                 pth, fn = os.path.split(keywords_fn)
                 fn, ext = os.path.splitext(fn)
@@ -403,18 +494,17 @@ class FVS(object):
                         , '{}.tre'.format(fn)
                         )
                 if not trees is None:
-                    
                     self.write_treelist(trees, trees_fn)
+                    self._artifacts.append(trees_fn)
 
-                elif not self.inv_trees is None:
-                    self.write_treelist(self.inv_trees, trees_fn)
-                
+                elif not self.inventory_trees is None:
+                    self.write_treelist(self.inventory_trees, trees_fn)
+                    self._artifacts.append(trees_fn)
+
                 else:
-                    log.warn('No inventory tree records, assume a bareground projection.')
-                    if not keywords.find('NOTREES'):
+                    if not os.path.exists(trees_fn) and not keywords.find('NOTREES'):
+                        log.info('No inventory tree records, assume a bareground projection.')
                         keywords += kw.NOTREES()
-                
-                self._artifacts.append(trees_fn)
 
         else:
             # Handle keywords as a file path
@@ -425,13 +515,13 @@ class FVS(object):
             msg = 'The keyword file does not exist: {}'.format(keywords_fn)
             log.error(msg)
             raise ValueError(msg)
-                
+
         # fvs_init requires a path
         r = self.fvs_step.fvs_init(keywords_fn)
 
         # TODO: Handle and format error codes
         if not r == 0:
-            log.error('FVS returned non-zero: {}'.format(r))
+            log.error(f'FVS returned non-zero: {r}\n{keywords_fn}')
 
         if self.stochastic:
             self.set_random_seed()
@@ -476,7 +566,7 @@ class FVS(object):
     def grow(self, cycles=1):
         """
         Execute the FVS projection for an arbitrary number of cycles.
-        
+
         Args:
             cycles: Number of cycles to execute. Set to zero to run all
                 remaining cycles.
@@ -499,6 +589,7 @@ class FVS(object):
         """
         Cleanup when the instance is destroyed.
         """
+        log.debug(f'Cleanup FVS() object: "{self._workspace}"')
         try:
             if self.fvs_step.sim_status > 0:
                 self.end_projection()
@@ -511,19 +602,17 @@ class FVS(object):
         Terminate the current projection.
         """
 
-        if self.fvs_step.sim_status <= 0:
-            # Nothing to do
-            return
-
-        # Finalize the projection
-        r = self.fvs_step.fvs_end()
+        r = 0
+        if self.fvs_step.sim_status > 0:
+            # Finalize the projection
+            r = self.fvs_step.fvs_end()
 
         if self.cleanup:
             for a in self._artifacts:
 #                 try: os.remove(a)
 #                 except: pass
                 try:
-                    if os.path.isdir(a):
+                    if os.path.exists(a) and os.path.isdir(a):
                         shutil.rmtree(a)
                         if a == self._workspace:
                             self._workspace = None
@@ -531,7 +620,8 @@ class FVS(object):
                         os.remove(a)
 
                 except:
-                    print(f'Unable to remove: {a}')
+                    log.debug(f'Unable to remove: {a}')
+                    # raise
                     pass
 
             self._artifacts = []
@@ -578,25 +668,29 @@ class FVS(object):
 
         tmplt = self.treelist_tmplt
 
+        for fld in self._tree_fields:
+            if not fld in trees.columns:
+                trees[fld] = self._inv_tree_defaults[fld]
+
         if hasattr(trees, 'fvs_treelist'):
             with open(treelist_path, 'w') as out:
                 out.write(trees.fvs_treelist())
 
         elif isinstance(trees, pd.DataFrame):
-            names = list(trees.columns)
-            try:
-                with open(treelist_path, 'w') as out:
-                    for r, rec in trees.iterrows():
-                        out.write(tmplt.format(
-                                **dict(zip(names, rec))) + '\n')
-            except KeyError:
-                print('Treelist template: {}'.format(tmplt))
-                raise
-            
-            except:
-                print(trees.info())
-                raise
-            
+            names = self._tree_fields
+
+            with open(treelist_path, 'w') as out:
+                for r, rec in trees.loc[:,names].iterrows():
+                    print(tmplt)
+                    print(trees.loc[:,names].info())
+                    # row = tmplt.format(**dict(zip(names, rec))) + '\n'
+                    row = tmplt.format(**rec.to_dict()) + '\n'
+                    print(row)
+                    out.write(row)
+
+            # print(trees.loc[:,names])
+            raise
+
         elif isinstance(trees, (list, tuple)):
             with open(treelist_path, 'w') as out:
                 for plot in trees:
@@ -635,13 +729,13 @@ class FVS(object):
     @property
     def num_cycles(self):
         return int(self.globals.ncyc)
-    
+
     # @property
     # def treelist(self):
     #     """
     #     Return a dataframe of FVS trees.
     #     """
-    #     cols = 
+    #     cols =
 
     @property
     def trees(self):
@@ -660,12 +754,12 @@ class FVS(object):
             n = self.tree_data.num_recs[i]
             for c in cols:
                 trees[c][m:m+n] = getattr(self.tree_data, c)[:n,i]
-            
+
             trees['cycle'][m:m+n] = i
             trees['year'][m:m+n] = self.globals.iy[i]
-            
+
             m += n
-            
+
         return pd.DataFrame(trees)
         # return trees
 
@@ -674,18 +768,18 @@ class FVS(object):
         """
         Return the FVS summary table as a Pandas DataFrame
         """
-        
+
         cols = ['year','age','tpa','tcuft','mcuft','mbdft'
                 ,'rem_tpa','rem_tcuft','rem_mcuft','rem_mbdft'
                 ,'baa','ccf','top_ht','per_len','accr','mort'
                 ,'sam_wgt','for_type','size','stock']
-        
+
         df = pd.DataFrame.from_records(
             self.fvslib.globals.iosum[:, :self.num_cycles + 1].T
             , columns=cols)
-        
+
         return df
-        
+
     def get_summary(self, variable):
         """
         Return an FVS summary value through the current projection cycle.
